# Optimize DeckLink capture latency - v1.6.0