… CHANGELOG.md
… THREAD_PROGRESS.md
… docs/DESIGN_PHILOSOPHY.md
… src/common/version.h