@zbynekdrlik 2025-07-16
## Compilation Fix Applied

Fixed the compilation error where `CaptureStatistics` was missing the `metadata` field.

### Changes:
- Added `std::unordered_map<std::string, std::string> metadata;` to `CaptureStatistics` struct in `ICaptureDevice.h`
- This allows the v1.6.0 low-latency statistics to be properly stored

### Build Status:
The code should now compile successfully. The macro redefinition warnings for `NDI_BRIDGE_VERSION_MINOR` are likely from CMake and can be ignored - they don't prevent compilation.

### Ready for Testing:
Please rebuild and test the v1.6.0 optimizations. Looking forward to seeing the latency improvements! 🚀