@zbynekdrlik 2025-07-17
## v1.6.2 - Critical Bug Fix Applied

Fixed the "No frames received" issue that was occurring in v1.6.1.

### Problem Found
- DeckLink was detecting the signal but the frame callback wasn't being invoked
- Logs showed 0 frames captured, 293 frames dropped over 5 seconds
- The `VideoInputFrameArrived` callback was never being called

### Root Cause
The frame callback was being set **BEFORE** the device was initialized:
```cpp
// WRONG ORDER (v1.6.1):
m_captureDevice->SetFrameCallback(...);  // Set callback on uninitialized device
m_captureDevice->Initialize(targetDevice);  // Initialize device
m_captureDevice->StartCapture();
```

### Fix Applied
Reordered the initialization sequence to ensure the device is ready before setting callbacks:
```cpp
// CORRECT ORDER (v1.6.2):
m_captureDevice->Initialize(targetDevice);  // Initialize device first
m_captureDevice->SetFrameCallback(...);     // Set callback on initialized device
m_captureDevice->StartCapture();
```

### Testing Required
Please rebuild and test v1.6.2. You should now see:
- Frames being received properly
- Zero-copy counter incrementing
- "TRUE ZERO-COPY: UYVY direct to NDI" log message
- No more "No frames received" errors

The performance should match the expected v1.6.1 goals once the frames are flowing correctly.