@zbynekdrlik 2025-07-17
## v1.6.3 - Second Callback Fix Applied

Found and fixed a second initialization order issue in the AppController.

### Additional Problem Found
Even after fixing the DeckLink callback initialization in v1.6.2, frames were still being dropped because:
- AppController was setting its frame callbacks **AFTER** calling `startCapture()`
- By the time the callbacks were set, DeckLink was already capturing and dropping frames

### Complete Fix Applied
Two initialization order fixes were needed:

1. **v1.6.2 (DeckLink level)**:
   ```cpp
   // Fixed order in DeckLinkCapture::startCapture():
   Initialize() → SetFrameCallback() → StartCapture()
   ```

2. **v1.6.3 (AppController level)**:
   ```cpp
   // Fixed order in AppController::initialize():
   setFrameCallback() → setErrorCallback() → startCapture()
   ```

### Callback Chain (Now Complete)
The full callback chain is now properly initialized:
1. AppController sets callbacks on DeckLinkCapture
2. DeckLinkCapture initializes DeckLinkCaptureDevice
3. DeckLinkCapture sets callback on DeckLinkCaptureDevice
4. DeckLinkCaptureDevice starts capture
5. Frames flow: DeckLink SDK → DeckLinkCaptureCallback → DeckLinkCaptureDevice → DeckLinkCapture → AppController

### Testing Required
Please rebuild and test v1.6.3. With both fixes in place, you should now see:
- Frames being received and processed
- Zero-copy counter incrementing for UYVY format
- TRUE zero-copy performance as intended
- No more frame drops due to missing callbacks

This completes the callback initialization fixes. The DeckLink implementation should now achieve the same low-latency performance as the Linux V4L2 implementation.